pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sharpness-aware-pc"
version = "0.1.0"
description = "Hessian-trace regularization for probabilistic circuits."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "Sahil Sidheekh" }]
dependencies = [
    "torch>=2.1",
    "torchvision>=0.16",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "pyjuice>=2.2.0",
    "pyro-ppl>=1.9.1",
    "numpy>=1.24",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "networkx>=3.1",
    "opt-einsum>=3.3",
    "scikit-learn>=1.3",
    "scipy>=1.10",
    "tqdm>=4.65",
    "wandb>=0.16",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.12",
    "ruff>=0.6",
]

[tool.setuptools.packages.find]
include = ["hessian_reg*", "packages*"]

[tool.ruff]
line-length = 100
extend-exclude = [
    "packages",
]
